Add files via upload
The project's central aim is to determine which physicochemical properties contribute to categorizing wine as 'good' or 'not good.' This dataset contains information about red and white wine variants of the Portuguese "Vinho Verde" wine, focusing solely on physicochemical and sensory variables while omitting factors like grape type or brand. The classification task involves differentiating between 'good' and 'not good' wines, where 'good' wines have a quality score of 7 or higher, while the rest are considered 'not good.' This project will explore which physicochemical characteristics play a significant role in classifying wine quality, and it may involve feature selection techniques to identify the most relevant attributes for this classification.